[CD-2845]: Binding ArtifactStream to Service and add migration (#7175)
* Add artifactStreamIds to Service

* Add ArtifactStreamServiceBindingService

* Extend ArtifactStreamServiceBindingService with OwnedByArtifactStream

* Fix ArtifactStreamServiceBindingServiceImpl

* Fix unit tests

* Use ServiceKeys instead of string literals

* Add payload to POST request

* Add artifact stream binding to artifact stream resource

* Replace some usages of ArtifactStream::serviceId with Service::artifactStreamIds

* Replace some more usages of ArtifactStream::serviceId with Service::artifactStreamIds

* Fix style warnings

* Fix few tests

* Fix errors

* Fix some unit tests

* Update functional test utils

* Add description in ServiceGenerator

* Add migration

* Fix styling

* Update migration to not use ArtifactStreamKeys.serviceId

* Fix cpd checks

* Rename ArtifactStreamServiceBindingServiceTest

* Fix errors

+++ 71-rest/src/main/java/software/wings/resources/ArtifactStreamResource.java
@@ -18,6 +18,7 @@ import software.wings.security.annotations.AuthRule;
 import software.wings.security.annotations.Scope;
 import software.wings.service.intfc.AppService;
 import software.wings.service.intfc.ArtifactStreamService;
+import software.wings.service.intfc.ArtifactStreamServiceBindingService;
 
 import java.util.List;
 import java.util.Map;
@@ -46,17 +47,21 @@ import javax.ws.rs.QueryParam;
 @AuthRule(permissionType = SERVICE, skipAuth = true)
 public class ArtifactStreamResource {
   private ArtifactStreamService artifactStreamService;
+  private ArtifactStreamServiceBindingService artifactStreamServiceBindingService;
   private AppService appService;
 
   /**
    * Instantiates a new Artifact stream resource.
    *
-   * @param artifactStreamService the artifact stream service
-   * @param appService            the app service
+   * @param artifactStreamService               the artifact stream service
+   * @param artifactStreamServiceBindingService the artifact stream service binding service
+   * @param appService                          the app service
    */
   @Inject
-  public ArtifactStreamResource(ArtifactStreamService artifactStreamService, AppService appService) {
+  public ArtifactStreamResource(ArtifactStreamService artifactStreamService,
+      ArtifactStreamServiceBindingService artifactStreamServiceBindingService, AppService appService) {
     this.artifactStreamService = artifactStreamService;
+    this.artifactStreamServiceBindingService = artifactStreamServiceBindingService;
     this.appService = appService;
   }
 
@@ -67,6 +72,16 @@ public class ArtifactStreamResource {
    */
   public void setArtifactStreamService(ArtifactStreamService artifactStreamService) {
     this.artifactStreamService = artifactStreamService;
+  }
+
+  /**
+   * Sets artifact stream service binding service.
+   *
+   * @param artifactStreamServiceBindingService the artifact stream service binding service
+   */
+  public void setArtifactStreamServiceBindingService(
+      ArtifactStreamServiceBindingService artifactStreamServiceBindingService) {
+    this.artifactStreamServiceBindingService = artifactStreamServiceBindingService;
   }
 
   /**
@@ -123,8 +138,19 @@ public class ArtifactStreamResource {
     if (!appService.exist(appId)) {
       throw new NotFoundException("application with id " + appId + " not found.");
     }
+
     artifactStream.setAppId(appId);
-    return new RestResponse<>(artifactStreamService.create(artifactStream));
+    // NOTE: artifactStream and binding must be created atomically
+    ArtifactStream savedArtifactStream = artifactStreamService.create(artifactStream);
+    try {
+      artifactStreamServiceBindingService.create(
+          appId, savedArtifactStream.getServiceId(), savedArtifactStream.getUuid());
+    } catch (Exception e) {
+      artifactStreamService.delete(appId, savedArtifactStream.getUuid());
+      throw e;
+    }
+
+    return new RestResponse<>(savedArtifactStream);
   }
 
   /**
@@ -143,6 +169,7 @@ public class ArtifactStreamResource {
       @QueryParam("appId") String appId, @PathParam("streamId") String streamId, ArtifactStream artifactStream) {
     artifactStream.setUuid(streamId);
     artifactStream.setAppId(appId);
+    // NOTE: should not update serviceId
     return new RestResponse<>(artifactStreamService.update(artifactStream));
   }
 
@@ -158,7 +185,20 @@ public class ArtifactStreamResource {
   @Timed
   @ExceptionMetered
   public RestResponse delete(@QueryParam("appId") String appId, @PathParam("id") String id) {
+    ArtifactStream artifactStream = artifactStreamService.get(appId, id);
+    if (artifactStream == null) {
+      return new RestResponse<>();
+    }
+
+    // NOTE: artifactStream and binding must be deleted atomically
     artifactStreamService.delete(appId, id);
+    try {
+      artifactStreamServiceBindingService.delete(appId, artifactStream.getServiceId(), id);
+    } catch (Exception e) {
+      artifactStreamService.create(artifactStream);
+      throw e;
+    }
+
     return new RestResponse<>();
   }
